Use 32-bit integers like the specification states (#28)
* Change i64 to i32
* Change u64 to u32
* Change f64 to f32
* Bump version

+++ src/v1_6/messages/cancel_reservation.rs
@@ -4,7 +4,7 @@ use crate::v1_6::types::CancelReservationStatus;
 #[serde(rename_all = "camelCase")]
 pub struct CancelReservationRequest {
     /// Required. Id of the reservation to cancel.
-    pub reservation_id: i64,
+    pub reservation_id: i32,
 }
 
 #[derive(serde::Serialize, serde::Deserialize, Debug, Clone, PartialEq, Default)]
