+++ src/v1_6/messages/get_diagnostics.rs
@@ -9,10 +9,10 @@ pub struct GetDiagnosticsRequest {
     pub location: String,
     /// Optional. This specifies how many times Charge Point must try to upload the diagnostics before giving up. If this field is not present, it is left to Charge Point to decide how many times it wants to retry.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub retries: Option<i64>,
+    pub retries: Option<i32>,
     /// Optional. The interval in seconds after which a retry may be attempted. If this field is not present, it is left to Charge Point to decide how long to wait between attempts.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub retry_interval: Option<i64>,
+    pub retry_interval: Option<i32>,
     /// Optional. This contains the date and time of the oldest logging information to include in the diagnostics.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub start_time: Option<DateTime<Utc>>,
