+++ src/v1_6/messages/update_firmware.rs
@@ -9,12 +9,12 @@ pub struct UpdateFirmwareRequest {
     pub location: String,
     /// Optional. This specifies how many times Charge Point must try to download the firmware before giving up. If this field is not present, it is left to Charge Point to decide how many times it wants to retry.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub retries: Option<i64>,
+    pub retries: Option<i32>,
     /// Required. This contains the date and time after which the Charge Point is allowed to retrieve the (new) firmware.
     pub retrieve_date: DateTime<Utc>,
     /// Optional. The interval in seconds after which a retry may be attempted. If this field is not present, it is left to Charge Point to decide how long to wait between attempts.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub retry_interval: Option<i64>,
+    pub retry_interval: Option<i32>,
 }
 
 /// This contains the field definition of the UpdateFirmware.conf PDU sent by the Charge Point to the Central System in response to a UpdateFirmware.req PDU. See also Update Firmware
