+++ src/v2_0_1/datatypes/set_monitoring_data_type.rs
@@ -8,10 +8,10 @@ use crate::v2_0_1::enumerations::monitor_enum_type::MonitorEnumType;
 #[serde(rename_all = "camelCase")]
 pub struct SetMonitoringDataType {
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub id: Option<i64>,
+    pub id: Option<i32>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub transaction: Option<bool>,
-    pub value: f64,
+    pub value: f32,
     #[serde(rename = "type")]
     pub kind: MonitorEnumType,
     pub severity: u8,
