+++ src/v2_0_1/messages/get_display_message.rs
@@ -10,9 +10,9 @@ use crate::v2_0_1::enumerations::message_state_enum_type::MessageStateEnumType;
 pub struct GetDisplayMessagesRequest {
     /// If provided the Charging Station shall returnDisplay Messages of the given ids. This field SHALL NOTcontain more ids than set inNumberOfDisplayMessages.maxLimit
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub id: Option<Vec<i64>>,
+    pub id: Option<Vec<i32>>,
     /// The Id of this request
-    pub request_id: i64,
+    pub request_id: i32,
     /// If provided the Charging Station shall returnDisplay Messages with the given priority only
     #[serde(skip_serializing_if = "Option::is_none")]
     pub priority: Option<MessagePriorityEnumType>,
